+++ R/custom_fetches.R
@@ -90,13 +90,23 @@ NULL
 
 #' @describeIn custom_fetches Method for ranks of bggGames.
 #'
-.fetch_ranks_games <- function(xml)
+.fetch_ranks_gms <- function(xml)
     .fetch_ranks(xml, "statistics/ratings/ranks")
 
-#' @describeIn custom_fetches Method for bestplayers of bggCollection.
+#' @describeIn custom_fetches Method for ranks of bggCollection.
 #'
 .fetch_ranks_cllctn <- function(xml)
     .fetch_ranks(xml, "stats/rating/ranks")
+
+#' @describeIn custom_fetches Method for families of bggGames.
+#'
+.fetch_families_gms <- function(xml)
+    .fetch_families(xml, "statistics/ratings/ranks")
+
+#' @describeIn custom_fetches Method for families of bggCollection.
+#'
+.fetch_families_cllctn <- function(xml)
+    .fetch_families(xml, "stats/rating/ranks")
 
 #' @describeIn custom_fetches Method for bestplayers of bggGames.
 #'
@@ -130,6 +140,16 @@ NULL
     res <- split(res$outcome, factor(res$i, levels = seq_along(xml)))
     res <- unname(res)
     return(res)
+}
+
+.fetch_families <- function(xml, families_xpath)
+{
+    xpath <- paste0(families_xpath, "/rank[@type = 'family']")
+    fams <- lapply(xml, xml_find_all, xpath = xpath)
+    fams <- lapply(fams, function(x)
+        str_remove(xml_attr(x, attr = "friendlyname"),
+                   pattern = " Rank$"))
+    return(fams)
 }
 
 .fetch_ranks <- function(xml, ranks_xpath)