+++ NEWS.md
@@ -1,3 +1,8 @@
+# bggAnalytics 0.2.0
+
+* Slight optimization of variable fetching.
+* Added 'ranks' and 'families' variables to bggGames and bggCollection classes.
+
 # bggAnalytics 0.1.1
 
 * Fixed a bug which caused fetch variables to have improper names.